refactor(core): move metrics tracker into utils module
Relocate createMetricsTracker from src/core/src/metrics.ts to src/core/src/utils/metrics.ts because it is a general-purpose utility rather than a top-level core domain module.

Updated Core and utils export surfaces plus internal tests so imports resolve from the new logical location while preserving behavior.
