+++ src/core/test/reporting-metrics.test.ts
@@ -1,7 +1,7 @@
 import assert from "node:assert/strict";
 import test from "node:test";
 
-import { createMetricsTracker } from "../src/metrics.js";
+import { createMetricsTracker } from "../src/utils/metrics.js";
 
 void test("snapshot exposes accumulated metrics as plain objects", () => {
     const tracker = createMetricsTracker({ category: "demo" });
